Render beaches on results page 🏖
